+++ Network/Telegram/API/Bot/Property/Accessible.hs
@@ -6,7 +6,7 @@ import "lens" Control.Lens (Lens')
 
 import Network.Telegram.API.Bot.Object (Object)
 import Network.Telegram.API.Bot.Object.Update.Callback (Callback (Datatext))
-import Network.Telegram.API.Bot.Object.Update.Message (Message (Direct, Forward))
+import Network.Telegram.API.Bot.Object.Update.Message (Message (Direct, Forward, Reply))
 import Network.Telegram.API.Bot.Object.Update.Message.Content (Content)
 import Network.Telegram.API.Bot.Object.Update.Message.Origin (Origin)
 
@@ -16,6 +16,7 @@ class Object source => Accessible target source where
 instance Accessible Content Message where
 	access f (Direct msg_id origin content) = (\content' -> Direct msg_id origin content') <$> f content
 	access f (Forward msg_id origin content) = (\content' -> Forward msg_id origin content') <$> f content
+	access f (Reply msg_id origin content msg) = (\content' -> Reply msg_id origin content' msg) <$> f content
 
 instance Accessible Origin Callback where
 	access f (Datatext cq_id msg dttxt) = flip
@@ -24,3 +25,4 @@ instance Accessible Origin Callback where
 instance Accessible Origin Message where
 	access f (Direct msg_id origin content) = (\origin' -> Direct msg_id origin' content) <$> f origin
 	access f (Forward msg_id origin content) = (\origin' -> Forward msg_id origin' content) <$> f origin
+	access f (Reply msg_id origin content msg) = (\origin' -> Reply msg_id origin' content msg) <$> f origin